// src/types.d.ts
export type FormData = {
    Geschlecht: string;
    Ferientage: number;
    'Hygiene Level': number;
    Transportmittel: string;
    Unterkunftsart: string;
    'Eigene spezifische Artikel': string;
    Jahreszeit: string;
    Reiseart: string; 
}

export type packList = {
    'Underwear': number;
    'Socks': number;
    'T-Shirts': number;
    'Sweatshirts': number;
    'Pants': number;
    'Shorts': number;
    'Hats': number;
    'Caps': number;
    'Shoes Light': number;
    'Shoes Heavy': number;
    'Hiking Shoes': number;
    'Bathing Pants': number;
    'Skis': number;
    'Ski Boots': number;
    'Ski Jacket': number;
    'Gas Cooker': number;
    'Food Help': number;
    'Sleeping Bag': number;
    'Toothbrush': number;
    'Toothpaste': number;
    'Soap': number;
    'Deodorant': number;
    'Shaver': number;
    'Shaving Creme': number;
    'Towel': number;
    'Shampoo': number;
    'Hand Creme': number;
    'Lip Balm': number;
    'Perfume': number;
    'Nail Clippers': number;
    'Cotton Swabs': number;
    'Cleanser': number;
    'Moisturizer': number;
    'Face Cream': number;
    'Bug Spray': number;
    'Bras': number;
    'Book Bags': number;
    'Suitcases': number;
    'Small Bags': number;
}